feat(Additional feature): :sparkles: added saving new and existing pages to the scan result page
added initialization of database during load up of Scan Results page for storing Results Data. New tables made for patients and results. Storing of images as BLOB is yet to be implemented and sharing data for results among pages is not fully implemented. For now, a  class object is used.