feat(Create Function): :sparkles: redefined tables and successful storage of images in the Results table
results of a scanned image are now saved to the results table along with the images (orig, preprocessed, and Grad-CAM) together with the corresponding patient. Saving to an existing patient and viewing of records yet to be implemented. When saving new image, the fields of the save to new patient are not left out. It should be erased once a record is updated.