feat: Added additional messages for results
added a message for the results screen to provide suggestions for the patient's next time in regards to the results of their scan. also made small changes to formatting to make screen look less wonky. Added feature to login text inputs so that tab can be pressed without it adding 2 spaces, also logs in when pressing enter.